Css fixed, validation added, loadingbutton used
