+++ src/components/loadingbutton/LoadingButton.tsx
@@ -4,20 +4,14 @@ import React, { FC } from 'react'
 
 interface ButtonProps {
   title: string
-  onClick: () => Promise<void>
+  onClick: () => void
   loading: boolean
+  type?: 'button' | 'submit' | 'reset'
 }
 
-const LoadingButton: FC<ButtonProps> = ({ title, onClick, loading }) => {
+const LoadingButton: FC<ButtonProps> = ({ title, onClick, loading, type }) => {
   return (
-    <Button
-      onClick={() => {
-        onClick().finally(() => {})
-      }}
-      disabled={loading}
-      variant="contained"
-      color="primary"
-    >
+    <Button onClick={onClick} disabled={loading} variant="contained" color="primary" type={type}>
       {loading ? <CircularProgress size={24} color="inherit" /> : title}
     </Button>
   )
